refactor: aprimora arquitetura de notas e corrige falhas de segurança
- implementa o padrão "Find or Create" no service de Matchups, permitindo a criação de notas apenas com os nomes dos campeões
- reestrutura a hierarquia de Schemas do Pydantic (NoteBase, NoteCreate, NoteResponse) para garantir a entrada correta de dados via JSON
- refatora os Routers de Notas e Perfis para consumirem os schemas adequadamente, eliminando Query Parameters incorretos
- corrige a instanciação da entidade Note, garantindo que o SQLAlchemy gerencie as chaves primárias
- implementa verificação de autorização (posse do perfil) nos métodos de update e delete de notas, e trata-os como uma exceção específica
- remove a redundância na entidade Notes, que buscava o user_id e profile_id

+++ services/notes_services.py
@@ -1,6 +1,6 @@
 from repositories import profiles_repository, notes_repository
-from services import profiles_services
-from core.exceptions import ProfileNotFound, NotesNotFound, NoteNotFound
+from services import profiles_services, matchups_service
+from core.exceptions import ProfileNotFound, NotesNotFound, NoteNotFound, MatchupNotFound, Unauthorized
 from models.notes_models import Note
 
 def get_notes_by_profile_service(db, user_id, nickname, tagline):
@@ -12,24 +12,31 @@ def get_notes_by_profile_service(db, user_id, nickname, tagline):
         raise NotesNotFound
     return notes
     
-def create_note_service(db, user_id, nickname, tagline, content):
+def create_note_service(db, user_id, nickname, tagline, player_champion_name, enemy_champion_name, content):
     profile=profiles_services.auth_profile_service(db, user_id, nickname, tagline)
+    matchup=matchups_service.get_matchup_by_names(db, player_champion_name, enemy_champion_name)
+    if not matchup:
+        raise MatchupNotFound
     note=Note(
-        user_id=user_id,
-        content=content,
-        profile_id=profile.id)
+        matchup_id=matchup.id,
+        profile_id=profile.id,
+        content=content)
     return notes_repository.create_note_repo(db, note)
 
 def update_note_service(db, user_id, nickname, tagline, note_id, new_content):
-    profiles_services.auth_profile_service(db, user_id, nickname, tagline)
+    profile=profiles_services.auth_profile_service(db, user_id, nickname, tagline)
     note=notes_repository.get_note_by_note_id(db, note_id)
     if not note:
         raise NoteNotFound
-    return notes_repository.update_note_repo(db, note, new_content)
+    if note.profile_id == profile.id:
+        return notes_repository.update_note_repo(db, note, new_content)
+    raise Unauthorized
 
 def delete_note_service(db, user_id, nickname, tagline, note_id):
-    profiles_services.auth_profile_service(db, user_id, nickname, tagline)
+    profile=profiles_services.auth_profile_service(db, user_id, nickname, tagline)
     note=notes_repository.get_note_by_note_id(db, note_id)
     if not note:
         raise NoteNotFound
-    return notes_repository.delete_note_repo(db, note)
+    if note.profile_id == profile.id:
+        return notes_repository.delete_note_repo(db, note)
+    raise Unauthorized
